moved parse WWW-Authenticate to new header package.
